Fix broken home page due to constraint.
Adding the constraint to `/:id` causes '/' to also match to
`pools#show`, but the `:id` param is empty. What we'd like it to match
is a string that is not 'websocket' but also not empty, and I couldn't
figure out a regex that does that.  What I did here works around that
problem by making the websocket route higher priority, but I'm not sure
if it's the best solution.